Changed NR_TsENS_VX to X2
The X2 port is physically closer to the other ports for the CNR4 sensor. This keeps CNR4 wiring confined to one sector of the CNR4 wiring panel. In case original CNR4 cables from Kipp & Zonen are used, a resistor must be installed on the X port and must reach the neighboring port