rH multiplier documentation
Added information from the manual on the rH multiplier for Vaisala HMP155 sensor. Depending on desired output, (percentage or fraction) a different multiplier us uses. Currently, the HMP155 sensor for WSF is set to report rH as percentage, hence mult = 0.1, even tough the comment in the code says to use 0.001 (i.e. fraction).